Move DiscoveryImageConfigForm from common to cim/components/modals
